+++ Common/Include/pointgrey.h
@@ -35,8 +35,8 @@ namespace TopGear
         int currentFormatIndex = -1;
     public:
         virtual int Flip(bool vertical, bool horizontal) override;
-        virtual int GetExposure(uint16_t& val) override;
-        virtual int SetExposure(uint16_t val) override;
+        virtual int GetExposure(uint32_t& val) override;
+        virtual int SetExposure(uint32_t val) override;
         virtual int GetGain(uint16_t& gainR, uint16_t& gainG, uint16_t& gainB) override;
         virtual int SetGain(uint16_t gainR, uint16_t gainG, uint16_t gainB) override;
         explicit PointGrey(std::shared_ptr<IVideoStream> &vs);
